refactor to package; rename modules
