+++ setup.py
@@ -1,0 +1,13 @@
+from setuptools import setup
+
+setup(
+    name='mooshimeter',
+    version='1.0',
+    packages=['mooshimeter', 'mooshimeter.vendor'],
+    url='',
+    license='GPLv3',
+    author='James Whong',
+    author_email='',
+    description='',
+    install_requires=['pyserial']
+)
